requirements.txt
# Sales Analytics API v2 Dependencies

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL driver for production

# HTTP Client
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
